requirements.txt
fastapi
uvicorn
openai
requests
pynacl
discord.py